Moved out the 'simple-templates-query' part of the repo to a separate repo
This part of the tree has become a project on its own right, it is not really an experiment any more. So I decided to create a separate github repo (http://github.com/iherman/CSVPlus) and continue any kind of development there. The tree in the current repo is hereby frozen; the README.md file has been changed accordingly.