Simplify CSV2RDF row processing.
Previously, there were separate steps to find unique subjects for a row, and then process each cell having the same subject. However, it is much simpler to just simply process each cell, which results in exactly the same triples. It does mean that certain statements may be emitted multiple times, but this is perfectly acceptable in the abstract syntax, and results in only a single triple in a resulting Graph. I believe simplified process this allows is well worth the percieved efficiency (and complication) of doing subject-at-a-time processing.